requirements.txt
streamlit==1.28.1
sqlalchemy==2.0.23
pandas==2.1.4
python-dotenv==1.0.0
email-validator==2.1.0
phonenumbers==8.13.26